third_party/WebKit/LayoutTests/bluetooth/getPrimaryServices/get-same-service.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('TwoHeartRateServicesAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => Promise.all([
      gattServer.getPrimaryServices('heart_rate'),
      gattServer.getPrimaryServices('heart_rate')]))
    .then(services => {
      let services1 = services[0];
      let services2 = services[1];
      assert_equals(services1.length, services2.length);
      // getPrimaryServices should return the same object if it was created
      // earlier. https://crbug.com/495270
      // TODO(ortuno): Change to assert_equals.
      for (let i = 1; i < services1.length; i++) {
        assert_not_equals(services1[i], services2[i],
                          'Should return the same service as the first call.');
      }
    });
}, 'Calls to get the same service should return the same object.');
</script>
